adjust remove/delete image by property

+++ class/v1.0/Property.php
@@ -137,8 +137,8 @@ class Property{
         $result = 0;
         $varField = $type == "" ? "" : "request";
         $varTable = $type == "" ? "" : "_request";
-        //$this->remove_image($token, $path); //remove image before
-        //$this->delete_data_image($token); //delete data image
+        $this->remove_image($token, $type, $path); //remove image before
+        $this->delete_data_image($token, $type); //delete data image
 
         $text = "DELETE FROM $this->table$varTable WHERE id_produk$varField = '$token'";
         $query = mysql_query($text);
@@ -161,10 +161,12 @@ class Property{
         return $result;
     }
 
-    public function delete_data_image($token){
+    public function delete_data_image($token, $type){
         $result = 0;
+        $varField = $type == "" ? "" : "request";
+        $varTable = $type == "" ? "" : "_request";
         
-        $text = "DELETE FROM t_property_image WHERE pi_property = '$token'";
+        $text = "DELETE FROM photo_produk$varTable WHERE id_produk$varField = '$token'";
         $query = mysql_query($text);
         if(mysql_affected_rows() == 1){
             $result = 1;
@@ -184,28 +186,21 @@ class Property{
         return $result;
     }
 
-    public function remove_image($token, $path){
+    public function remove_image($token, $type, $path){
         $result = 0;
-        $flag_img = 0;
-        $flag_img_thmb = 0;
+        $varField = $type == "" ? "" : "request";
+        $varTable = $type == "" ? "" : "_request";
 
-        $text = "SELECT pi_img, pi_img_thmb FROM t_property_image WHERE pi_property = '$token'";
+        $text = "SELECT id_photo, file_photo FROM photo_produk$varTable WHERE id_produk$varField = '$token'";
         $query = mysql_query($text);
         while($row = mysql_fetch_array($query, MYSQL_ASSOC)){
-            $deleteImg = $path.$row['pi_img'];
-            if(file_exists($deleteImg)){
-                unlink($deleteImg);
-                $flag_img = 1;
-            }
-
-            $deleteImgThmb = $path.$row['pi_img_thmb'];
-            if(file_exists($deleteImgThmb)){
-                unlink($deleteImgThmb);
-                $flag_img_thmb = 1;
-            }
-
-            if($flag_img == 1 && $flag_img_thmb ==1){
-                $result = 1;
+            $value = $row['file_photo'];
+            if($value != ""){
+                $deleteImg = $path."produk_photo/".$value;
+                if (file_exists($deleteImg)) {
+                    unlink($deleteImg);
+                    $result = 1;
+                }
             }
         }
         return $result;
